refactor: restore module boundaries (I/O out of model/presentation)
0.12.0 architectural cleanup — model/ and presentation/ are pure DTO
layers again; nothing imports crate::prelude internally or crosses the
application->storage edge.

- HttpClient + make_http_request moved model/http.rs -> application/http.rs
  (re-exported so the public path is unchanged)
- build/extract_market_hierarchy moved model/utils.rs -> application/
  market_hierarchy.rs; model/utils.rs deleted
- Session / WebsocketInfo moved application/auth.rs -> model/auth.rs (plain
  data types), re-exported from application::auth for path compatibility
- DatabaseConfig moved into application/config.rs (storage::config
  re-exports it); Config::pg_pool() removed (was application->sqlx) —
  callers use storage::utils::create_connection_pool
- the five From<&ItemUpdate> impls moved out of presentation/* into
  application/streaming_convert.rs; presentation DTOs expose pure
  from_fields constructors and no longer import lightstreamer_rs
- nine files switched from crate::prelude to direct module paths

Boundary greps all empty (prelude / reqwest+sqlx+lightstreamer in
model+presentation / storage in application+model). semver passes at
0.12.0. Public paths preserved via re-exports.
