+++ examples/other/src/bin/tx_loop.rs
@@ -29,14 +29,14 @@ async fn main() -> Result<(), ig_client::error::AppError> {
     debug!("Loaded config: database={}", config.database);
 
     // Build the Postgres pool once at startup
-    let pool = match config.pg_pool().await {
+    let pool = match create_connection_pool(&config.database).await {
         Ok(pool) => {
             info!("Postgres pool established");
             pool
         }
         Err(e) => {
             error!("Failed to establish database connection: {}", e);
-            return Err(e.into());
+            return Err(e);
         }
     };
