+++ src/application/mod.rs
@@ -6,7 +6,13 @@ pub mod client;
 pub mod config;
 /// Dynamic market streamer with thread-safe subscription management
 pub mod dynamic_streamer;
+/// HTTP client and request execution with rate limiting and finite retry
+pub mod http;
 /// Service interfaces and traits
 pub mod interfaces;
+/// Market navigation hierarchy traversal helpers
+pub mod market_hierarchy;
 /// Rate limiter module for API request throttling
 pub mod rate_limiter;
+/// Adapters from Lightstreamer `ItemUpdate` to presentation-layer DTOs
+pub mod streaming_convert;
