+++ src/presentation/trade.rs
@@ -1,6 +1,5 @@
 use crate::presentation::order::{Direction, OrderType, Status, TimeInForce};
 use crate::presentation::serialization::{option_string_empty_as_none, string_as_float_opt};
-use lightstreamer_rs::subscription::ItemUpdate;
 use pretty_simple_display::{DebugPretty, DisplaySimple};
 use serde::{Deserialize, Serialize};
 use serde_json;
@@ -175,38 +174,35 @@ pub struct WorkingOrderUpdate {
 }
 
 impl TradeData {
-    /// Converts a Lightstreamer ItemUpdate to a TradeData object
+    /// Builds a [`TradeData`] from pre-extracted streaming fields.
+    ///
+    /// This is transport-agnostic: it takes plain field maps rather than a
+    /// Lightstreamer `ItemUpdate`, so the presentation layer carries no
+    /// dependency on the streaming transport. The `ItemUpdate` adapter lives in
+    /// [`crate::application::streaming_convert`].
     ///
     /// # Arguments
-    ///
-    /// * `item_update` - The ItemUpdate from Lightstreamer containing trade data
+    /// * `item_name` - Subscription item name (`None` when subscribed by position)
+    /// * `item_pos` - 1-based position of the item in the subscription
+    /// * `is_snapshot` - Whether this update is a snapshot
+    /// * `fields` - Current field values for the item
+    /// * `changed_fields` - Field values that changed in this update
     ///
     /// # Returns
-    ///
     /// A Result containing either the parsed TradeData or an error message
-    pub fn from_item_update(item_update: &ItemUpdate) -> Result<Self, String> {
-        // Extract the item_name, defaulting to an empty string if None
-        let item_name = item_update.item_name.clone().unwrap_or_default();
-
-        // Convert item_pos from usize to i32
-        let item_pos = item_update.item_pos as i32;
-
-        // Extract is_snapshot
-        let is_snapshot = item_update.is_snapshot;
-
-        // Convert fields
-        let fields = Self::create_trade_fields(&item_update.fields)?;
-
-        // Convert changed_fields by first creating a HashMap<String, Option<String>>
-        let mut changed_fields_map: HashMap<String, Option<String>> = HashMap::new();
-        for (key, value) in &item_update.changed_fields {
-            changed_fields_map.insert(key.clone(), Some(value.clone()));
-        }
-        let changed_fields = Self::create_trade_fields(&changed_fields_map)?;
+    pub fn from_fields(
+        item_name: Option<&str>,
+        item_pos: usize,
+        is_snapshot: bool,
+        fields: &HashMap<String, Option<String>>,
+        changed_fields: &HashMap<String, Option<String>>,
+    ) -> Result<Self, String> {
+        let fields = Self::create_trade_fields(fields)?;
+        let changed_fields = Self::create_trade_fields(changed_fields)?;
 
         Ok(TradeData {
-            item_name,
-            item_pos,
+            item_name: item_name.unwrap_or_default().to_string(),
+            item_pos: item_pos as i32,
             fields,
             changed_fields,
             is_snapshot,
@@ -262,34 +258,9 @@ impl TradeData {
     }
 }
 
-impl From<&ItemUpdate> for TradeData {
-    fn from(item_update: &ItemUpdate) -> Self {
-        Self::from_item_update(item_update).unwrap_or_default()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    /// Builds an `ItemUpdate` carrying a single TRADE field, mirroring how the
-    /// Lightstreamer TRADE subscription delivers OPU / WOU / CONFIRMS payloads.
-    ///
-    /// `item_name`, `item_pos` and `is_snapshot` are set to distinctive,
-    /// obviously-synthetic values so tests can assert whether they survive the
-    /// parse (they are dropped on the silent-default path — see the malformed
-    /// test).
-    fn trade_item_update(field_key: &str, field_value: &str) -> ItemUpdate {
-        let mut fields = HashMap::new();
-        fields.insert(field_key.to_string(), Some(field_value.to_string()));
-        ItemUpdate {
-            item_name: Some("TRADE:SYNTHETIC".to_string()),
-            item_pos: 2,
-            fields,
-            changed_fields: HashMap::new(),
-            is_snapshot: true,
-        }
-    }
 
     // Realistic, sanitized OPU payload. Field names mirror IG's Open Position
     // Update stream (camelCase); the deal ids are obviously synthetic and carry
@@ -310,91 +281,6 @@ mod tests {
         "expiry": "DFB"
     }"#;
 
-    // Realistic, sanitized WOU payload. Field names mirror IG's Working Order
-    // Update stream (camelCase); the deal ids are obviously synthetic.
-    const SANITIZED_WOU_JSON: &str = r#"{
-        "dealReference": "REF_SYNTHETIC_0002",
-        "dealId": "DIAAAAF00SYNTH02",
-        "direction": "SELL",
-        "epic": "CS.D.EURUSD.CFD.IP",
-        "status": "OPEN",
-        "dealStatus": "ACCEPTED",
-        "level": "1.1000",
-        "size": "10000",
-        "currency": "USD",
-        "timestamp": "2024-01-15T10:31:00.000",
-        "channel": "PublicRestOTC",
-        "expiry": "-",
-        "stopDistance": "10.5",
-        "limitDistance": "20.0",
-        "guaranteedStop": false,
-        "orderType": "LIMIT",
-        "timeInForce": "GOOD_TILL_CANCELLED",
-        "goodTillDate": ""
-    }"#;
-
-    #[test]
-    fn test_from_item_update_parses_open_position_update() {
-        let item = trade_item_update("OPU", SANITIZED_OPU_JSON);
-
-        let result = TradeData::from_item_update(&item);
-        assert!(result.is_ok(), "OPU payload should parse: {result:?}");
-        let data = result.expect("OPU parse checked ok above");
-
-        // Update metadata is preserved on the success path.
-        assert_eq!(data.item_name, "TRADE:SYNTHETIC");
-        assert_eq!(data.item_pos, 2);
-        assert!(data.is_snapshot);
-
-        let opu = data.fields.opu.expect("OPU should be populated");
-        assert_eq!(opu.deal_reference.as_deref(), Some("REF_SYNTHETIC_0001"));
-        assert_eq!(opu.deal_id.as_deref(), Some("DIAAAAF00SYNTH01"));
-        assert_eq!(opu.deal_id_origin.as_deref(), Some("DIAAAAF00SYNTH01"));
-        assert_eq!(opu.direction, Some(Direction::Buy));
-        assert_eq!(opu.epic.as_deref(), Some("IX.D.DAX.DAILY.IP"));
-        assert_eq!(opu.status, Some(Status::Open));
-        assert_eq!(opu.deal_status, Some(Status::Accepted));
-        assert_eq!(opu.level, Some(18000.5));
-        assert_eq!(opu.size, Some(1.0));
-        assert_eq!(opu.currency.as_deref(), Some("EUR"));
-        assert_eq!(opu.expiry.as_deref(), Some("DFB"));
-
-        // No WOU or CONFIRMS present in this update.
-        assert!(data.fields.wou.is_none());
-        assert!(data.fields.confirms.is_none());
-    }
-
-    #[test]
-    fn test_from_item_update_parses_working_order_update() {
-        let item = trade_item_update("WOU", SANITIZED_WOU_JSON);
-
-        let result = TradeData::from_item_update(&item);
-        assert!(result.is_ok(), "WOU payload should parse: {result:?}");
-        let data = result.expect("WOU parse checked ok above");
-
-        let wou = data.fields.wou.expect("WOU should be populated");
-        assert_eq!(wou.deal_reference.as_deref(), Some("REF_SYNTHETIC_0002"));
-        assert_eq!(wou.deal_id.as_deref(), Some("DIAAAAF00SYNTH02"));
-        assert_eq!(wou.direction, Some(Direction::Sell));
-        assert_eq!(wou.epic.as_deref(), Some("CS.D.EURUSD.CFD.IP"));
-        assert_eq!(wou.status, Some(Status::Open));
-        assert_eq!(wou.deal_status, Some(Status::Accepted));
-        assert_eq!(wou.level, Some(1.1000));
-        assert_eq!(wou.size, Some(10000.0));
-        assert_eq!(wou.currency.as_deref(), Some("USD"));
-        assert_eq!(wou.stop_distance, Some(10.5));
-        assert_eq!(wou.limit_distance, Some(20.0));
-        assert_eq!(wou.guaranteed_stop, Some(false));
-        assert_eq!(wou.order_type, Some(OrderType::Limit));
-        assert_eq!(wou.time_in_force, Some(TimeInForce::GoodTillCancelled));
-        // Empty string good-till-date degrades to None.
-        assert!(wou.good_till_date.is_none());
-
-        // No OPU or CONFIRMS present in this update.
-        assert!(data.fields.opu.is_none());
-        assert!(data.fields.confirms.is_none());
-    }
-
     #[test]
     fn test_create_trade_fields_parses_opu_json() {
         // Exercise the JSON-parse helper directly (independent of ItemUpdate).
@@ -410,49 +296,6 @@ mod tests {
         let opu = fields.opu.expect("OPU should be populated");
         assert_eq!(opu.deal_id.as_deref(), Some("DIAAAAF00SYNTH01"));
         assert_eq!(opu.level, Some(18000.5));
-    }
-
-    #[test]
-    fn test_from_item_update_malformed_opu_is_error_on_direct_path() {
-        let item = trade_item_update("OPU", "{ this is not valid json ");
-
-        // The direct parse path surfaces the failure as an Err.
-        let result = TradeData::from_item_update(&item);
-        assert!(
-            result.is_err(),
-            "malformed OPU JSON must surface an error on the direct parse path"
-        );
-    }
-
-    #[test]
-    fn test_from_impl_swallows_malformed_opu_into_default() {
-        // KNOWN GAP (pinned, do not "fix" here): `From<&ItemUpdate>` funnels any
-        // parse failure through `.unwrap_or_default()`, so a malformed TRADE
-        // payload is silently swallowed into `TradeData::default()` instead of
-        // being surfaced. This test PINS that current behavior so a future change
-        // to make the swallow deliberate (e.g. logging, or propagating the error)
-        // is a conscious, reviewed decision rather than an accidental regression.
-        //
-        // TODO(streaming): decide whether `From<&ItemUpdate>` should log the
-        // dropped payload at WARN or expose the parse error instead of defaulting.
-        let item = trade_item_update("OPU", "{ this is not valid json ");
-
-        let data = TradeData::from(&item);
-
-        // Everything degrades to the Default, including the update metadata that
-        // was present on the source ItemUpdate (item_name / item_pos / snapshot).
-        assert!(
-            data.item_name.is_empty(),
-            "silent-default path discards item_name"
-        );
-        assert_eq!(data.item_pos, 0, "silent-default path discards item_pos");
-        assert!(
-            !data.is_snapshot,
-            "silent-default path discards is_snapshot"
-        );
-        assert!(data.fields.opu.is_none());
-        assert!(data.fields.wou.is_none());
-        assert!(data.fields.confirms.is_none());
     }
 
     #[test]
