+++ src/storage/market_database.rs
@@ -1,4 +1,4 @@
-use crate::prelude::{MarketData, MarketNode};
+use crate::presentation::market::{MarketData, MarketNode};
 use crate::storage::market_persistence::{MarketHierarchyNode, MarketInstrument};
 use chrono::{DateTime, Utc};
 use sqlx::{AssertSqlSafe, Executor, PgPool, Row};
